pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
authors = [
    {name = "Adel Napier",email = "adelnapier@ya.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PyYAML"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.uv]
dev-dependencies = [
    "ruff",
    "pytest",
    "pytest-cov"
]


[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
read-json = "gendiff.scripts.read_json_files:main"


[tool.ruff]
line-length = 88
select = ["E", "F", "C", "N", "Q"]
ignore = ["E501"]


[tool.hatch.build.targets.wheel]
packages = ["gendiff"]
